LingTeX-Word: no Single anywhere -- the stage-1 Overflow, diagnosed
RunDocTests compiled (the seven declaration fixes held) and died in its first
section with run-time error 6, Overflow, at a call passing the literal 0 into a
"ByVal wantSpaceAfter As Single" parameter -- no arithmetic anywhere near it.

That is the second time. Stage 1's undiagnosed failure was "s1 = 10" into a
Single, in a different module, and it "went away" when the procedure was split up
-- recorded then as a workaround, with an honest note that it was not a diagnosis.
Two occurrences of one shape on the same Mac build is the diagnosis: converting an
integer into a Single fails on this build, depending on the stack frame. TypeCheck
passed every type including Single because it tests each in a one-line function,
the one kind of frame where it never fails; that is also why "make it smaller"
looked like a fix.

So the engine uses Double throughout: 96 lines across seven modules, every width,
size, gap and position. Double is VBA's native floating type and what the
JavaScript reference has used all along. ReadSingle is ReadDouble. CSng is CDbl.
Line counts are unchanged, so nothing else moved.

Two PROVEN modules are among the seven -- modWrap and modTests -- which is why
RunAllTests must be run again after this: its 79 checks re-prove them. Keeping
Single at that boundary was the alternative, and it would have kept the very type
that fails in play, with the frame-dependence deciding when. The Single in
TypeCheck and MicroDiagnose stays deliberately: it is the reproduction, and the
comment there now says what came of it.

QUICKSTART's "unexplained failure" section is rewritten as the explained one, with
the rule that falls out: error 6 with no arithmetic in sight means look for a
Single first.

vba-lint ALL PASS, parity 78 passed.

+++ LingTeX-Word/src/modSettings.bas
@@ -22,12 +22,12 @@ Private Const VAR_PREFIX As String = "LingTeX_"
 '-- Defaults ------------------------------------------------------------------
 ' Horizontal air between alignment columns, in points.  Enough to read the
 ' columns as separate without the example sprawling.
-Private Const DEF_GAP As Single = 6
+Private Const DEF_GAP As Double = 6
 ' Vertical air between wrap lines, in points.
-Private Const DEF_LINE_GAP As Single = 6
+Private Const DEF_LINE_GAP As Double = 6
 ' Indent applied to every wrap line after the first, in points.  Zero by default:
 ' a flush-left continuation keeps the columns of long examples comparable.
-Private Const DEF_CONT_INDENT As Single = 0
+Private Const DEF_CONT_INDENT As Double = 0
 ' What replaces a space found inside an interlinear cell.
 Private Const DEF_SPACE_REPL As String = "."
 ' Lowercase grammatical glosses so Word's small caps can render them.
@@ -43,16 +43,16 @@ Private Const DEF_REWRAP_ON_SELECTION As Boolean = False
 ' -- TYPED GETTERS ----------------------------------------------------------
 '=============================================================================
 
-Public Function SettingGap(doc As Document) As Single
-    SettingGap = ReadSingle(doc, "Gap", DEF_GAP)
+Public Function SettingGap(doc As Document) As Double
+    SettingGap = ReadDouble(doc, "Gap", DEF_GAP)
 End Function
 
-Public Function SettingLineGap(doc As Document) As Single
-    SettingLineGap = ReadSingle(doc, "LineGap", DEF_LINE_GAP)
+Public Function SettingLineGap(doc As Document) As Double
+    SettingLineGap = ReadDouble(doc, "LineGap", DEF_LINE_GAP)
 End Function
 
-Public Function SettingContIndent(doc As Document) As Single
-    SettingContIndent = ReadSingle(doc, "ContIndent", DEF_CONT_INDENT)
+Public Function SettingContIndent(doc As Document) As Double
+    SettingContIndent = ReadDouble(doc, "ContIndent", DEF_CONT_INDENT)
 End Function
 
 Public Function SettingSpaceReplacement(doc As Document) As String
@@ -101,15 +101,15 @@ End Function
 ' -- TYPED SETTERS ----------------------------------------------------------
 '=============================================================================
 
-Public Sub SetSettingGap(doc As Document, ByVal v As Single)
+Public Sub SetSettingGap(doc As Document, ByVal v As Double)
     WriteVar doc, "Gap", CStr(v)
 End Sub
 
-Public Sub SetSettingLineGap(doc As Document, ByVal v As Single)
+Public Sub SetSettingLineGap(doc As Document, ByVal v As Double)
     WriteVar doc, "LineGap", CStr(v)
 End Sub
 
-Public Sub SetSettingContIndent(doc As Document, ByVal v As Single)
+Public Sub SetSettingContIndent(doc As Document, ByVal v As Double)
     WriteVar doc, "ContIndent", CStr(v)
 End Sub
 
@@ -157,18 +157,18 @@ Private Function ReadString(doc As Document, ByVal nm As String, _
     On Error GoTo 0
 End Function
 
-Private Function ReadSingle(doc As Document, ByVal nm As String, _
-        ByVal dflt As Single) As Single
+Private Function ReadDouble(doc As Document, ByVal nm As String, _
+        ByVal dflt As Double) As Double
     Dim s As String
-    ReadSingle = dflt
+    ReadDouble = dflt
     s = ReadString(doc, nm, "")
     If s = "" Then Exit Function
     On Error Resume Next
-    ReadSingle = CSng(s)
-    If Err.Number <> 0 Then ReadSingle = dflt
+    ReadDouble = CDbl(s)
+    If Err.Number <> 0 Then ReadDouble = dflt
     Err.Clear
     On Error GoTo 0
-    If ReadSingle < 0 Then ReadSingle = 0
+    If ReadDouble < 0 Then ReadDouble = 0
 End Function
 
 Private Function ReadBool(doc As Document, ByVal nm As String, _
